refactor(csv): extract pure generation helpers
